Moved "MqttConnectorManager" into a different class", adjusted init.py and other dependencies accordingly (1 file 1 class solution)
